# python/examples/rectangle/server/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@bytesforlife.com
    name: bytes for life GmbH
    url: https://www.bytesforlife.com
  description: representation is a minimalistic cross-language client-server framework.
    representations are client-side references to server-side objects. they can be
    interacted with as if they were local objects. the source code is found on https://github.com/bytesforlife/representation.
  license:
    name: MIT License
    url: https://github.com/bytesforlife/representation/blob/main/LICENSE.txt
  title: representation API
  version: 1.0.0
externalDocs:
  description: technical documentation
  url: https://bytesforlife.github.io/representation/
servers:
- url: /
paths:
  /representation/{representation_uuid}:
    delete:
      description: release the representation of a server-side object
      operationId: release_representation
      parameters:
      - explode: false
        in: path
        name: representation_uuid
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: release a representation
      x-openapi-router-controller: openapi_server.controllers.default_controller
    get:
      description: download a serialized version of a server-side object itself (instead
        of its representation). think of this as a de-referencing operation, downloading
        the object to the client.
      operationId: download_representation
      parameters:
      - explode: false
        in: path
        name: representation_uuid
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: download a serialized version of a server-side object
      x-openapi-router-controller: openapi_server.controllers.default_controller
    patch:
      description: call a server-side object such as a static function, a member function,
        or in case of a class, its constructor
      operationId: call_representation
      parameters:
      - explode: false
        in: path
        name: representation_uuid
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ArgsKwargs'
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: call a server-side object
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /representation/{representation_uuid}/{attribute_name}:
    get:
      description: create a representation of an attribute of a representation
      operationId: create_attribute
      parameters:
      - explode: false
        in: path
        name: representation_uuid
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: attribute_name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: create a representation of an attribute of a representation
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /representations:
    get:
      description: returns a list of server-side objects that are available to be
        represented client-side. typical choices are classes, static functions, and
        other singletons. instances, their attributes, and other variables may also
        be referenced.
      operationId: list_representations
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: list available server-side objects
      x-openapi-router-controller: openapi_server.controllers.default_controller
    post:
      description: create a representation, a client-side reference to a server-side
        object. the representation will be wrapped client-side, so that it may be
        interacted with as if it were a local object.
      operationId: create_representation
      requestBody:
        $ref: '#/components/requestBodies/ArgsKwargs'
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: create a representation
      x-openapi-router-controller: openapi_server.controllers.default_controller
    put:
      description: upload an object to the server, and return a reference for future
        interaction.
      operationId: upload_representation
      requestBody:
        $ref: '#/components/requestBodies/ArgsKwargs'
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                properties:
                  type:
                    type: string
                type: object
          description: Success
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_default'
          description: Error
      summary: upload an object, and return its representation
      x-openapi-router-controller: openapi_server.controllers.default_controller
components:
  requestBodies:
    ArgsKwargs:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArgsKwargs'
      description: an ordered list of unnamed arguments, followed by a key-value dictionary
        of named arguments
      required: true
  responses:
    ResponseOk:
      content:
        application/json:
          schema:
            additionalProperties: true
            properties:
              type:
                type: string
            type: object
      description: Success
    ResponseError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_response_default'
      description: Error
  schemas:
    ArgsKwargs:
      additionalProperties: false
      example:
        args:
        - ""
        - ""
        kwargs:
          key: '{}'
      properties:
        args:
          items: {}
          type: array
        kwargs:
          additionalProperties: true
          type: object
      required:
      - args
      - kwargs
      type: object
    inline_response_default:
      properties:
        code:
          title: code
          type: integer
        message:
          title: message
          type: string
        stacktrace:
          title: stacktrace
          type: string
      required:
      - code
      - message
      title: inline_response_default
      type: object